Portfolio CRUD almost ready in new format
